+++ CHANGELOG.md
@@ -73,6 +73,7 @@ Here is an overview of all new **experimental** features:
 - **CPU Memory Scaler** Store forgotten logger ([#4022](https://github.com/kedacore/keda/issues/4022))
   **Kafka Scaler**: Support 0 in activationLagThreshold configuration ([#4137](https://github.com/kedacore/keda/issues/4137))
 - **Prometheus Metrics**: Expose Prometheus Metrics also when getting ScaledObject state ([#4075](https://github.com/kedacore/keda/issues/4075))
+- **Cassandra Scaler:** Fixed the assignment of port information ([4110])(https://github.com/kedacore/keda/issues/4110)
 
 
 ### Deprecations